+++ src/helpers/runCommand.ts
@@ -5,9 +5,13 @@ import type { CLIFlags } from '@eswatch/types'
 type RunOutput = (options: CLIFlags) => ChildProcess
 
 const runCommand: RunOutput = options => {
-  const commandToRun = typeof options.run === 'string' ? options.run : `node ${options.outdir}`
+  const commandToRun =
+    typeof options.run === 'string'
+      ? options.run
+      : `node --experimental-specifier-resolution=node ${options.outdir}`
   const [commandName] = commandToRun.split(' ')
   const commandParameters = commandToRun.split(' ').slice(1)
+  // Consider replacing this `spawn` with a `fork`
   const childProcess = spawn(commandName, commandParameters, { stdio: 'inherit' })
   return childProcess
 }